Move labeled-response capture out of send.c into label.c; split m_ack.c
send.c had grown by 550 lines of capture machinery that only touched it
at two points. It now lives in ircd/label.c with its interface in
include/label.h:

- send_buffer() calls label_capture_intercept(), which returns whether
  it took the line into the active capture.
- sendcmdto_one_hunted() asks label_capture_active_for() for the
  requester's active capture instead of reading the window statics.

struct MsgTagCtx and msgtagctx_init() were private to send.c; label.c
snapshots the context by value for every captured line, so they move
into send.h. The intercept takes the effective tag context (cache ctx or
explicit ctx) rather than the cache, so struct TagSendCache stays
private.

ms_ack() moves from m_batch.c into its own m_ack.c, mirroring the
one-handler-per-file layout, with the same CAP_BATCH + CAP_LABELED_
RESPONSE delivery gate as ms_batch() and parse.c's capture start.

No behaviour change. Comment pointers that said "in send.c" now say
label.c.
